+++ spec/spec_helper.rb
@@ -1,8 +1,11 @@
 # frozen_string_literal: true
 
+require "vcr"
 require "vcr/unused_cassettes"
 
 VCR::UnusedCassettes.eager_load!
+
+Dir[File.expand_path("support/**/*.rb", __dir__)].sort.each { |f| require f }
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure